refactor: clean up src/ directory structure
- Consolidate context/ into contexts/ (merge duplicate dirs)
- Rename Labs/ to labs/ (feature flags, not components)
- Update all import paths (29 files)
