+++ src/components/Labs/LabsButton.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { useLabsStore } from '../../store/labs';
-import { getFeature } from '../../Labs/features';
+import { getFeature } from '../../labs/features';
 import { SparklesIcon } from './icons';
 
 export function LabsButton() {
