+++ src/hooks/useFeatureFlag.ts
@@ -1,5 +1,5 @@
 import { useLabsStore } from '../store/labs';
-import type { FeatureId } from '../Labs/features';
+import type { FeatureId } from '../labs/features';
 
 export function useFeatureFlag(featureId: FeatureId): boolean {
   return useLabsStore((state) => state.isFeatureEnabled(featureId));
