+++ src/hooks/useInteraction.ts
@@ -2,7 +2,7 @@ import { useEffect, useLayoutEffect, useCallback, useRef, useMemo } from 'react'
 import type { RefObject } from 'react';
 import type { Coord, ResizeHandle } from '../types';
 import { useLayoutStore, useUndoableAction, useSelectionStore, useInteractionStore } from '../store';
-import { useMutations } from '../context/MutationsContext';
+import { useMutations } from '../contexts/MutationsContext';
 import { useGridCoords } from './useGridCoords';
 import { useCollabPresence } from './useCollabPresence';
 import { throttleRAF, cancelThrottledRAF } from '../utils/throttle';
